@@ -6,12 +6,14 @@
     "homepage": "https://github.com/jbouzekri/ConfigKnpMenuBundle",
     "license": "MIT",
     "require": {
+        "php": ">=7.4",
         "knplabs/knp-menu-bundle": "~3.0",
-        "symfony/event-dispatcher": "^4.0|^5.0",
-        "symfony/yaml": "^4.0|^5.0",
-        "symfony/security-core": "^4.0|^5.0"
+        "symfony/event-dispatcher": "^5.0|^6.0",
+        "symfony/yaml": "^5.0|^6.0",
+        "symfony/security-core": "^5.0|^6.0"
     },
     "require-dev": {
+        "roave/security-advisories": "dev-latest",
         "phpunit/phpunit": "^8.0",
         "squizlabs/php_codesniffer": "^3.0"
     },
