@@ -2,27 +2,30 @@
 
 namespace Jb\Bundle\PhumborBundle\Tests\DependencyInjection;
 
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Jb\Bundle\ConfigKnpMenuBundle\DependencyInjection\JbConfigKnpMenuExtension;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBag;
+use Jb\Bundle\ConfigKnpMenuBundle\Tests\DependencyInjection\Fixtures\Bundle1\JbTest1Bundle;
+use Jb\Bundle\ConfigKnpMenuBundle\Tests\DependencyInjection\Fixtures\Bundle2\JbTest2Bundle;
 
 /**
  * Test Extension
  *
  * @author Jonathan Bouzekri <jonathan.bouzekri@gmail.com>
  */
-class JbConfigKnpMenuExtensionTest extends \PHPUnit\Framework\TestCase
+class JbConfigKnpMenuExtensionTest extends TestCase
 {
     /**
      * Test loading data from file
      */
-    public function testLoading()
+    public function testLoading(): void
     {
         $menuConfiguration = self::loadConfiguration();
 
-        $this->assertEquals($menuConfiguration['main']['tree']['second_item']['label'], 'Second Item Label');
-        $this->assertEquals($menuConfiguration['main']['tree']['first_item']['label'], 'First Item Label');
-        $this->assertEquals(count($menuConfiguration['main']['tree']['second_item']['children']), 1);
+        $this->assertEquals('Second Item Label', $menuConfiguration['main']['tree']['second_item']['label']);
+        $this->assertEquals('First Item Label', $menuConfiguration['main']['tree']['first_item']['label']);
+        $this->assertCount(1, $menuConfiguration['main']['tree']['second_item']['children']);
     }
 
     public static function loadConfiguration()
@@ -39,20 +42,18 @@ class JbConfigKnpMenuExtensionTest extends \PHPUnit\Framework\TestCase
      *
      * @param array $data
      *
-     * @return \Symfony\Component\DependencyInjection\ContainerBuilder
+     * @return ContainerBuilder
      */
-    protected static function createContainer($data = array())
+    protected static function createContainer(array $data = array()): ContainerBuilder
     {
-        $container = new ContainerBuilder(new ParameterBag(array_merge(array(
+        return new ContainerBuilder(new ParameterBag(array_merge(array(
             'kernel.bundles'     => array(
                 'JbTest1Bundle' =>
-                    'Jb\\Bundle\\ConfigKnpMenuBundle\\Tests\\DependencyInjection\\Fixtures\\Bundle1\\JbTest1Bundle',
+                    JbTest1Bundle::class,
                 'JbTest2Bundle' =>
-                    'Jb\\Bundle\\ConfigKnpMenuBundle\\Tests\\DependencyInjection\\Fixtures\\Bundle2\\JbTest2Bundle'
+                    JbTest2Bundle::class
             ),
             'kernel.root_dir' => 'app'
         ), $data)));
-
-        return $container;
     }
 }
